requirements.txt
pandas
requests
dateparser
skpy
numpy
python-decouple